refactor(gui): fold Maintenance page builder into MaintenanceMixin
Extend MaintenanceMixin with the Tools-tab page builder and the
shared _make_action_row factory, pairing the builder with the slot
handlers that drive it. Both lived on WinpodxWindow; pulling them
into the same mixin keeps the Maintenance tab a single self-contained
unit and lets main_window.py shrink further.

Methods now on MaintenanceMixin: _build_maintenance_page (Pod
Management + System + Windows Update sections), _make_action_row
(icon + label + desc + handler row factory), plus the existing slot
handlers (_on_cleanup, _on_suspend, _on_resume, _on_open_desktop,
_on_timesync, _on_debloat, _on_apply_fixes, _on_enable_updates,
_on_disable_updates, _refresh_update_status).

No behaviour change; tests: 1096 passed.

main_window.py: 627 -> 411 lines (-216). Cumulative: 2745 -> 411 (-85%).
